piece tests removed from project, replaced by tetromino tests.
